chore(sequencer): separate test utils from shared benchmark and test utils (#1613)
## Summary
Separated utils only used by tests from utils used by both benchmarks
and tests.

## Background
This is to get rid of many strings of nested `expect`s and `allow`s
mentioned in #1585

## Changes
- Separated `astria-sequencer::test_utils.rs` into `test_utils.rs` and
`benchmark_and_test_utils.rs`.
- Separated `astria-sequencer::app::test_utils.rs` into
`app::test_utils.rs` and `app::benchmark_and_test_utils.rs`.
- Fixed all the imports which broke as a result.

## Testing
All tests working

## Related Issues
closes #1585
