+++ crates/astria-sequencer/src/app/tests_block_ordering.rs
@@ -24,13 +24,17 @@ use tendermint::{
 
 use super::test_utils::get_alice_signing_key;
 use crate::{
-    app::test_utils::{
-        get_bob_signing_key,
-        get_judy_signing_key,
-        initialize_app_with_storage,
-        mock_balances,
-        mock_tx_cost,
-        MockTxBuilder,
+    app::{
+        benchmark_and_test_utils::{
+            initialize_app_with_storage,
+            mock_balances,
+            mock_tx_cost,
+        },
+        test_utils::{
+            get_bob_signing_key,
+            get_judy_signing_key,
+            MockTxBuilder,
+        },
     },
     proposal::commitment::generate_rollup_datas_commitment,
 };
@@ -138,7 +142,6 @@ async fn app_process_proposal_ordering_fail() {
         .process_proposal(process_proposal.clone(), storage.clone())
         .await
         .expect_err("expected ordering error");
-
     assert!(
         format!("{result:?}").contains("transactions have incorrect transaction group ordering"),
         "process proposal should fail due to incorrect ordering"
