+++ crates/astria-sequencer/src/app/tests_breaking_changes.rs
@@ -50,25 +50,29 @@ use tendermint::{
 };
 
 use crate::{
-    app::test_utils::{
-        default_genesis_accounts,
-        get_alice_signing_key,
-        get_bridge_signing_key,
-        initialize_app,
-        initialize_app_with_storage,
-        proto_genesis_state,
-        BOB_ADDRESS,
-        CAROL_ADDRESS,
+    app::{
+        benchmark_and_test_utils::{
+            default_genesis_accounts,
+            initialize_app_with_storage,
+            proto_genesis_state,
+            BOB_ADDRESS,
+            CAROL_ADDRESS,
+        },
+        test_utils::{
+            get_alice_signing_key,
+            get_bridge_signing_key,
+            initialize_app,
+        },
     },
     authority::StateReadExt as _,
-    bridge::StateWriteExt as _,
-    proposal::commitment::generate_rollup_datas_commitment,
-    test_utils::{
+    benchmark_and_test_utils::{
         astria_address,
         astria_address_from_hex_string,
         nria,
         ASTRIA_PREFIX,
     },
+    bridge::StateWriteExt as _,
+    proposal::commitment::generate_rollup_datas_commitment,
 };
 
 #[tokio::test]
@@ -197,7 +201,7 @@ async fn app_execute_transaction_with_every_action_snapshot() {
     // setup for ValidatorUpdate action
     let update = ValidatorUpdate {
         power: 100,
-        verification_key: crate::test_utils::verification_key(1),
+        verification_key: crate::benchmark_and_test_utils::verification_key(1),
     };
 
     let rollup_id = RollupId::from_unhashed_bytes(b"testchainid");
