+++ crates/astria-sequencer/src/bridge/state_ext.rs
@@ -333,7 +333,7 @@ mod tests {
     use cnidarium::StateDelta;
 
     use super::*;
-    use crate::test_utils::astria_address;
+    use crate::benchmark_and_test_utils::astria_address;
 
     fn asset_0() -> asset::Denom {
         "asset_0".parse().unwrap()
