+++ crates/astria-sequencer/src/mempool/mempool_state.rs
@@ -41,7 +41,7 @@ mod tests {
             StateReadExt as _,
             StateWriteExt as _,
         },
-        test_utils::{
+        benchmark_and_test_utils::{
             astria_address,
             nria,
         },
