+++ crates/astria-sequencer/src/service/consensus.rs
@@ -227,7 +227,7 @@ mod tests {
 
     use super::*;
     use crate::{
-        app::test_utils::{
+        app::benchmark_and_test_utils::{
             mock_balances,
             mock_tx_cost,
         },
@@ -242,7 +242,7 @@ mod tests {
                 RollupDataSubmission {
                     rollup_id: RollupId::from_unhashed_bytes(b"testchainid"),
                     data: Bytes::from_static(b"hello world"),
-                    fee_asset: crate::test_utils::nria().into(),
+                    fee_asset: crate::benchmark_and_test_utils::nria().into(),
                 }
                 .into(),
             ])
@@ -452,7 +452,8 @@ mod tests {
         let accounts = if let Some(funded_key) = funded_key {
             vec![astria_core::generated::protocol::genesis::v1::Account {
                 address: Some(
-                    crate::test_utils::astria_address(funded_key.address_bytes()).to_raw(),
+                    crate::benchmark_and_test_utils::astria_address(funded_key.address_bytes())
+                        .to_raw(),
                 ),
                 balance: Some(10u128.pow(19).into()),
             }]
@@ -460,7 +461,7 @@ mod tests {
             vec![]
         };
         let genesis_state = {
-            let mut state = crate::app::test_utils::proto_genesis_state();
+            let mut state = crate::app::benchmark_and_test_utils::proto_genesis_state();
             state.accounts = accounts;
             state
         }
