+++ crates/astria-sequencer/src/service/info/mod.rs
@@ -202,6 +202,7 @@ mod tests {
         },
         app::StateWriteExt as _,
         assets::StateWriteExt as _,
+        benchmark_and_test_utils::nria,
         fees::{
             StateReadExt as _,
             StateWriteExt as _,
@@ -226,8 +227,8 @@ mod tests {
             .unwrap();
 
         state.put_base_prefix("astria".to_string()).unwrap();
-        state.put_native_asset(crate::test_utils::nria()).unwrap();
-        state.put_ibc_asset(crate::test_utils::nria()).unwrap();
+        state.put_native_asset(nria()).unwrap();
+        state.put_ibc_asset(nria()).unwrap();
 
         let address = state
             .try_base_prefixed(&hex::decode("a034c743bed8f26cb8ee7b8db2230fd8347ae131").unwrap())
@@ -236,7 +237,7 @@ mod tests {
 
         let balance = 1000;
         state
-            .put_account_balance(&address, &crate::test_utils::nria(), balance)
+            .put_account_balance(&address, &nria(), balance)
             .unwrap();
         state.put_block_height(height).unwrap();
         storage.commit(state).await.unwrap();
@@ -263,7 +264,7 @@ mod tests {
         assert!(query_response.code.is_ok());
 
         let expected_balance = AssetBalance {
-            denom: crate::test_utils::nria().into(),
+            denom: nria().into(),
             balance,
         };
