+++ crates/astria-sequencer/src/service/mempool/tests.rs
@@ -13,6 +13,7 @@ use tendermint::{
 
 use crate::{
     app::{
+        benchmark_and_test_utils::genesis_state,
         test_utils::MockTxBuilder,
         App,
     },
@@ -31,7 +32,7 @@ async fn future_nonces_are_accepted() {
     let metrics = Box::leak(Box::new(Metrics::noop_metrics(&()).unwrap()));
     let mut mempool = Mempool::new(metrics, 100);
     let mut app = App::new(snapshot, mempool.clone(), metrics).await.unwrap();
-    let genesis_state = crate::app::test_utils::genesis_state();
+    let genesis_state = genesis_state();
 
     app.init_chain(storage.clone(), genesis_state, vec![], "test".to_string())
         .await
@@ -61,7 +62,7 @@ async fn rechecks_pass() {
     let metrics = Box::leak(Box::new(Metrics::noop_metrics(&()).unwrap()));
     let mut mempool = Mempool::new(metrics, 100);
     let mut app = App::new(snapshot, mempool.clone(), metrics).await.unwrap();
-    let genesis_state = crate::app::test_utils::genesis_state();
+    let genesis_state = genesis_state();
 
     app.init_chain(storage.clone(), genesis_state, vec![], "test".to_string())
         .await
@@ -99,7 +100,7 @@ async fn can_reinsert_after_recheck_fail() {
     let metrics = Box::leak(Box::new(Metrics::noop_metrics(&()).unwrap()));
     let mut mempool = Mempool::new(metrics, 100);
     let mut app = App::new(snapshot, mempool.clone(), metrics).await.unwrap();
-    let genesis_state = crate::app::test_utils::genesis_state();
+    let genesis_state = genesis_state();
 
     app.init_chain(storage.clone(), genesis_state, vec![], "test".to_string())
         .await
@@ -147,7 +148,7 @@ async fn recheck_adds_non_tracked_tx() {
     let metrics = Box::leak(Box::new(Metrics::noop_metrics(&()).unwrap()));
     let mut mempool = Mempool::new(metrics, 100);
     let mut app = App::new(snapshot, mempool.clone(), metrics).await.unwrap();
-    let genesis_state = crate::app::test_utils::genesis_state();
+    let genesis_state = genesis_state();
 
     app.init_chain(storage.clone(), genesis_state, vec![], "test".to_string())
         .await
